+++ odoo/local-src/pricelist_discount/tests/test_pricelist_discount.py
@@ -3,6 +3,8 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from odoo.tests.common import SavepointCase, post_install, at_install
+from odoo import fields
+from odoo.exceptions import ValidationError
 
 import logging
 
@@ -433,3 +435,295 @@ class TestPricelistDiscount(SavepointCase):
         self.assertEqual(
             self.discount_pricelist_id, self.sale.discount_pricelist_id
         )
+
+    def test_check_dates(self):
+        """ Test exceptions with promotion dates """
+
+        ProductSupplierinfo = self.env['product.supplierinfo']
+        ProductSupplierinfo.search(
+            [('product_tmpl_id', '=', self.p1.product_tmpl_id.id)]).unlink()
+
+        # Create the default price
+        ProductSupplierinfo.create({
+            'name': self.supplier.id,
+            'product_tmpl_id': self.p1.product_tmpl_id.id,
+        })
+
+        # Test to create a new default price
+        with self.assertRaises(ValidationError):
+            ProductSupplierinfo.create({
+                'name': self.supplier.id,
+                'product_tmpl_id': self.p1.product_tmpl_id.id,
+            })
+
+        # Test to create a promo without end date
+        with self.assertRaises(ValidationError):
+            ProductSupplierinfo.create({
+                'name': self.supplier.id,
+                'product_tmpl_id': self.p1.product_tmpl_id.id,
+                'date_start': fields.Date.from_string('2018-01-01')
+            })
+
+        # Test to create a promo without start date
+        with self.assertRaises(ValidationError):
+            ProductSupplierinfo.create({
+                'name': self.supplier.id,
+                'product_tmpl_id': self.p1.product_tmpl_id.id,
+                'date_end': fields.Date.from_string('2018-01-01')
+            })
+
+        # Promo 1 (2018-01-01 -> 2018-03-31)
+        ProductSupplierinfo.create({
+            'name': self.supplier.id,
+            'product_tmpl_id': self.p1.product_tmpl_id.id,
+            'date_start': fields.Date.from_string('2018-01-01'),
+            'date_end': fields.Date.from_string('2018-03-31'),
+            'discount_sale': 10,
+            'discount_purchase': 15,
+        })
+
+        # Promo 3 (2018-08-01 -> 2018-12-31)
+        ProductSupplierinfo.create({
+            'name': self.supplier.id,
+            'product_tmpl_id': self.p1.product_tmpl_id.id,
+            'date_start': fields.Date.from_string('2018-08-01'),
+            'date_end': fields.Date.from_string('2018-12-31'),
+            'discount_sale': 10,
+            'discount_purchase': 15,
+        })
+
+        # Promo 2 (2018-04-01 -> 2018-06-30)
+        ProductSupplierinfo.create({
+            'name': self.supplier.id,
+            'product_tmpl_id': self.p1.product_tmpl_id.id,
+            'date_start': fields.Date.from_string('2018-04-01'),
+            'date_end': fields.Date.from_string('2018-06-30'),
+            'discount_sale': 10,
+            'discount_purchase': 15,
+        })
+
+        # Test overlaps (2018-12-01 -> 2019-03-01) blocked by the promo 3
+        # (2018-08-01 -> 2018-12-31)
+        with self.assertRaises(ValidationError):
+            ProductSupplierinfo.create({
+                'name': self.supplier.id,
+                'product_tmpl_id': self.p1.product_tmpl_id.id,
+                'date_start': fields.Date.from_string('2018-12-01'),
+                'date_end': fields.Date.from_string('2019-03-01'),
+            })
+
+        # Test overlaps (2018-03-01 -> 2018-06-01) blocked by the promo 1
+        # (2018-01-01 -> 2018-03-31) and 2 (2018-04-01 -> 2018-06-30)
+        with self.assertRaises(ValidationError):
+            ProductSupplierinfo.create({
+                'name': self.supplier.id,
+                'product_tmpl_id': self.p1.product_tmpl_id.id,
+                'date_start': fields.Date.from_string('2018-03-01'),
+                'date_end': fields.Date.from_string('2018-06-01'),
+            })
+
+        # Test inverse date_start and date_end
+        with self.assertRaises(ValidationError):
+            ProductSupplierinfo.create({
+                'name': self.supplier.id,
+                'product_tmpl_id': self.p1.product_tmpl_id.id,
+                'date_start': fields.Date.from_string('2017-12-31'),
+                'date_end': fields.Date.from_string('2017-01-01'),
+            })
+
+        # Test overlaps with different min_qty
+        # Promo 1 (2018-01-01 -> 2018-03-31) with min_qty == 100
+        ProductSupplierinfo.create({
+            'name': self.supplier.id,
+            'product_tmpl_id': self.p1.product_tmpl_id.id,
+            'date_start': fields.Date.from_string('2018-01-01'),
+            'date_end': fields.Date.from_string('2018-03-31'),
+            'min_qty': 100,
+            'discount_sale': 10,
+            'discount_purchase': 20,
+        })
+
+        # Test overlaps with different min_qty_sale
+        # Promo 2 (2018-04-01 -> 2018-06-30) with min_qty_sale == 25
+        ProductSupplierinfo.create({
+            'name': self.supplier.id,
+            'product_tmpl_id': self.p1.product_tmpl_id.id,
+            'date_start': fields.Date.from_string('2018-04-01'),
+            'date_end': fields.Date.from_string('2018-06-30'),
+            'min_qty_sale': 25,
+            'discount_sale': 11.5,
+            'discount_purchase': 15,
+        })
+
+        # Test overlaps with different min_qty_sale
+        # Promo 2 (2018-04-01 -> 2018-06-30) with min_qty_sale == 50
+        ProductSupplierinfo.create({
+            'name': self.supplier.id,
+            'product_tmpl_id': self.p1.product_tmpl_id.id,
+            'date_start': fields.Date.from_string('2018-04-01'),
+            'date_end': fields.Date.from_string('2018-06-30'),
+            'min_qty_sale': 50,
+            'discount_sale': 14,
+            'discount_purchase': 15,
+        })
+
+        # Test overlaps with the same min_qty and min_qty_sale
+        with self.assertRaises(ValidationError):
+            ProductSupplierinfo.create({
+                'name': self.supplier.id,
+                'product_tmpl_id': self.p1.product_tmpl_id.id,
+                'date_start': fields.Date.from_string('2018-04-01'),
+                'date_end': fields.Date.from_string('2018-06-30'),
+                'min_qty_sale': 50,
+                'discount_sale': 14,
+                'discount_purchase': 15,
+            })
+
+    def test_select_seller(self):
+        """ Test the method _select_seller_for_sale
+        and _select_seller_for_sale.
+
+        Default price: 100€
+        Promo 1 (2018-01-01 -> 2018-03-31)
+            Min sale: 0 - min purchase: 0
+                * - 10% on sale
+                * - 15% on purchase
+            Min sale: 0 - min purchase: 100
+                * - 10% on sale
+                * - 20% on purchase
+
+        Promo 2 (2018-04-01 -> 2018-06-30)
+            Min sale: 0 - min purchase: 0
+                * - 11% on sale
+                * - 13% on purchase
+            Min sale: 25 - min purchase: 0
+                * - 11.5% on sale
+                * - 13% on purchase
+            Min sale: 50 - min purchase: 0
+                * - 14% on sale
+                * - 13% on purchase
+        Promo 3 (2018-08-01 -> 2018-12-31)
+            Min sale: 0 - min purchase: 0
+                * - 8% on sale
+                * - 10% on purchase
+        """
+
+        ProductSupplierinfo = self.env['product.supplierinfo']
+        ProductSupplierinfo.search(
+            [('product_tmpl_id', '=', self.p1.product_tmpl_id.id)]).unlink()
+
+        # Create the default price
+        default_promo = ProductSupplierinfo.create({
+            'name': self.supplier.id,
+            'product_tmpl_id': self.p1.product_tmpl_id.id,
+            'price': 100,
+        })
+
+        # Promo 1 (2018-01-01 -> 2018-03-31)
+        promo_1 = ProductSupplierinfo.create({
+            'name': self.supplier.id,
+            'product_tmpl_id': self.p1.product_tmpl_id.id,
+            'date_start': fields.Date.from_string('2018-01-01'),
+            'date_end': fields.Date.from_string('2018-03-31'),
+            'discount_sale': 10,
+            'discount_purchase': 15,
+            'price': 100,
+        })
+
+        # Promo 1 (2018-01-01 -> 2018-03-31) with min_qty == 100
+        promo_1_min_100 = ProductSupplierinfo.create({
+            'name': self.supplier.id,
+            'product_tmpl_id': self.p1.product_tmpl_id.id,
+            'date_start': fields.Date.from_string('2018-01-01'),
+            'date_end': fields.Date.from_string('2018-03-31'),
+            'min_qty': 100,
+            'discount_sale': 10,
+            'discount_purchase': 20,
+            'price': 100,
+        })
+
+        # Promo 2 (2018-04-01 -> 2018-06-30)
+        promo_2 = ProductSupplierinfo.create({
+            'name': self.supplier.id,
+            'product_tmpl_id': self.p1.product_tmpl_id.id,
+            'date_start': fields.Date.from_string('2018-04-01'),
+            'date_end': fields.Date.from_string('2018-06-30'),
+            'discount_sale': 11,
+            'discount_purchase': 13,
+            'price': 100,
+        })
+
+        # Promo 2 (2018-04-01 -> 2018-06-30) with min_qty_sale == 25
+        promo_2_min_sale_25 = ProductSupplierinfo.create({
+            'name': self.supplier.id,
+            'product_tmpl_id': self.p1.product_tmpl_id.id,
+            'date_start': fields.Date.from_string('2018-04-01'),
+            'date_end': fields.Date.from_string('2018-06-30'),
+            'min_qty_sale': 25,
+            'discount_sale': 11.5,
+            'discount_purchase': 13,
+            'price': 100,
+        })
+
+        # Promo 2 (2018-04-01 -> 2018-06-30) with min_qty_sale == 50
+        promo_2_min_sale_50 = ProductSupplierinfo.create({
+            'name': self.supplier.id,
+            'product_tmpl_id': self.p1.product_tmpl_id.id,
+            'date_start': fields.Date.from_string('2018-04-01'),
+            'date_end': fields.Date.from_string('2018-06-30'),
+            'min_qty_sale': 50,
+            'discount_sale': 14,
+            'discount_purchase': 13,
+            'price': 100,
+        })
+
+        # Promo 3 (2018-08-01 -> 2018-12-31)
+        promo_3 = ProductSupplierinfo.create({
+            'name': self.supplier.id,
+            'product_tmpl_id': self.p1.product_tmpl_id.id,
+            'date_start': fields.Date.from_string('2018-08-01'),
+            'date_end': fields.Date.from_string('2018-12-31'),
+            'discount_sale': 8,
+            'discount_purchase': 10,
+            'price': 100,
+        })
+
+        # Test default promo
+        promo = self.p1._select_seller(
+            partner_id=self.supplier, quantity=20, date='2019-01-01')
+        self.assertEqual(promo, default_promo)
+
+        # Test promo 1
+        promo = self.p1._select_seller(
+            partner_id=self.supplier, quantity=20, date='2018-01-01')
+        self.assertEqual(promo, promo_1)
+
+        # Test promo 2
+        promo = self.p1._select_seller_for_sale(
+            partner_id=self.supplier, quantity=20, date='2018-05-01')
+        self.assertEqual(promo, promo_2)
+
+        # Test promo 3
+        promo = self.p1._select_seller_for_sale(
+            partner_id=self.supplier, quantity=20, date='2018-12-31')
+        self.assertEqual(promo, promo_3)
+
+        # Test promo 1 with min (purchase) 100
+        promo = self.p1._select_seller(
+            partner_id=self.supplier, quantity=100, date='2018-01-01')
+        self.assertEqual(promo, promo_1_min_100)
+
+        # Test promo 2 with min (sale) 40
+        promo = self.p1._select_seller_for_sale(
+            partner_id=self.supplier, quantity=40, date='2018-05-01')
+        self.assertEqual(promo, promo_2_min_sale_25)
+
+        # Test promo 2 with min (sale) 120
+        promo = self.p1._select_seller_for_sale(
+            partner_id=self.supplier, quantity=120, date='2018-05-01')
+        self.assertEqual(promo, promo_2_min_sale_50)
+
+        # Test promo 2 with min (purchase) 40 (method _select_seller)
+        promo = self.p1._select_seller(
+            partner_id=self.supplier, quantity=40, date='2018-05-01')
+        self.assertEqual(promo, promo_2_min_sale_50)